Add Execution Noise to trading styles
Market makers to create transactions didn't work. They just reduced the spread arbitrarily small but no trades would be executed for long time and only because of the nosie orders. Made the price action flat for a long time. The noise has to be embedded in each order

this is kinda something i wanted to avoid because i feel it's a bit unrealistic

Also added Fundamentalists which have a internal variable holding what they believe the true value is. This is a randomly chosen number centred at market.p_f

This seems to help stabalise price action but again it feels slightly inelegant

Also removed the OHLC stuff from previous commit. That was a kinda crappy idea that introduced strange artifacts